+++ market/src/main/java/university/market/member/utils/JwtTokenProvider.java
@@ -1,52 +1,12 @@
 package university.market.member.utils;
+public interface JwtTokenProvider {
 
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import java.util.Date;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-import university.market.member.exception.MemberException;
-import university.market.member.exception.MemberExceptionType;
+    String generateToken(final String email);
 
-@Component
-public class JwtTokenProvider {
-    @Value("${security.jwt.secret-key}")
-    private String secretKey;
-    @Value("${security.jwt.expire-length}")
-    private int expireTimeMilliSecond;
+    String generateToken(final String email, final ExpireDateSupplier expireDateSupplier);
 
-    public String generateToken(final String email) {
-        final Date now = new Date();
+    void validateToken(String token);
 
-        return Jwts.builder()
-                .claim("email", email)
-                .expiration(new Date(now.getTime() + expireTimeMilliSecond))
-                .signWith(SignatureAlgorithm.HS256, secretKey)
-                .compact();
-    }
+    String extractEmail(final String token);
 
-    String generateToken(final String email, final ExpireDateSupplier expireDateSupplier) {
-        return Jwts.builder()
-                .claim("email", email)
-                .expiration(expireDateSupplier.expireDate())
-                .signWith(SignatureAlgorithm.HS256, secretKey)
-                .compact();
-    }
-
-
-    public String extractEmail(final String token) {
-        try {
-            return Jwts.parser()
-                    .setSigningKey(secretKey).build()
-                    .parseSignedClaims(token)
-                    .getBody()
-                    .get("email")
-                    .toString();
-        } catch (final ExpiredJwtException exception) {
-            throw new MemberException(MemberExceptionType.EXPIRED_ACCESS_TOKEN);
-        } catch (final Exception exception) {
-            throw new MemberException(MemberExceptionType.INVALID_ACCESS_TOKEN);
-        }
-    }
 }
